requirements.txt
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.7
aiofiles==23.2.0
aiosqlite==0.19.0
pytrends==4.9.2
pandas==2.1.4
numpy==1.26.4